pyproject.toml
[project]
name = "hyperbot"
version = "0.1.0"
description = "Trading bot for Hyperliquid with Telegram integration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.2",
    "uvicorn[standard]>=0.38.0",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.3",
    "loguru>=0.7.3",
    "hyperliquid-python-sdk>=0.20.0",
    "httpx>=0.28.1",
    "aiohttp>=3.13.2",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
    "python-telegram-bot>=22.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "mypy>=1.18.2",
    "black>=25.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "N",   # pep8-naming
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001", "ARG002", "ARG004", "ARG005"]  # Unused arguments in tests (fixtures)
"src/bot/**/*.py" = ["ARG001", "ARG004"]  # Telegram handler signature requirements
"src/api/main.py" = ["ARG001"]  # FastAPI lifespan signature requirement
".vulture_whitelist.py" = ["B018", "N801"]  # Vulture whitelist pattern (useless expressions, class naming)

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true

# Per-module options for third-party libraries
[[tool.mypy.overrides]]
module = [
    "hyperliquid.*",
    "telegram.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.4",
    "vulture>=2.14",
]
